Delete krypton config
Krypton was a cool bit of software that allowed SSH keys to be stored on
your phone in a secure enclave, but allowed computers to access it much
like an SSH agent. Unfortunately, the company behind it went under, the
code was acquired and turned into crypto-stealing malware. C'est la vie.